all done

+++ main.py
@@ -15,7 +15,8 @@ def logged_in_as_user():
         print("| Press 3 for Withdraw Money.          |")
         print("| Press 4 for Take Loan.               |")
         print("| Press 5 for Transaction History.     |")
-        print("| Press 6 for Logout.                  |")
+        print("| Press 6 for Profile Details.         |")
+        print("| Press 7 for Logout.                  |")
         print("|______________________________________|")
         
         choice = int(input("Enter your choice : "))
@@ -34,6 +35,8 @@ def logged_in_as_user():
         elif choice == 5:
             user.show_transaction_history()
         elif choice == 6:
+            user.profile_details()
+        elif choice == 7:
             print("Logging out...")
             sleep(random.randint(2, 3))
             break
@@ -61,10 +64,15 @@ def logged_in_as_admin():
             email = input("Enter email : ")
             phone = input("Enter phone no : ")
             password = input("Enter password : ")
-            bank.create_account(name, email, phone, password)
+            bank.create_account(name, email, phone, password, "user")
         elif choice == 2:
             account = int(input("Enter the account number you want to delete : "))
-            bank.delete_account(account)
+            conf = input("\nTo delete this account, type \"CONFIRM\" : ")
+            if conf == 'CONFIRM':
+                bank.delete_account(account)
+                print("Account deleted.")
+            else:
+                print("Text didn't match.")
         elif choice == 3:
             bank.show_users()
         elif choice == 4:
@@ -76,7 +84,12 @@ def logged_in_as_admin():
         elif choice == 7:
             bank.loan_off()
         elif choice == 8:
-            bank.declare_bankrupt()
+            bankrupt = input("\nTo declare bank as bankrupt, type \"BANKRUPT\" : ")
+            if bankrupt == 'BANKRUPT':
+                bank.declare_bankrupt()
+                print("Bank is declared as Bankrupt.")
+            else:
+                print("Text didn't match.")
         elif choice == 9:
             print("Logging out...")
             sleep(random.randint(2, 3))
@@ -88,43 +101,49 @@ while True:
     op = input("Register/Login/Exit (r/l/e) : ")
     if op == 'r':
         user_type = input("User/Admin (u/a) : ")
-        name = input("Enter your name : ")
-        email = input("Enter your email : ")
-        phone = input("Enter your phone no : ")
-        password = input("Enter a strong password : ")
         
-        if user_type == 'u':
-            user = Account(name, email, phone, password, "user")
-            print(f"Welcome {name} to Pocket Vari Bank. Your Account number is, {user.account_no}.")
-            bank.account_list.append(user)
-        elif user_type == 'a':
-            user = Account(name, email, phone, password, "admin")
-            print(f"Welcome {name} to Pocket Vari Bank. Your Administrative Account number is, {user.account_no}.")
-            bank.account_list.append(user)
+        if user_type == 'a' or user_type == 'u':
+            name = input("Enter your name : ")
+            email = input("Enter your email : ")
+            phone = input("Enter your phone no : ")
+            password = input("Enter a strong password : ")
+            
+            if user_type == 'u':
+                user = Account(name, email, phone, password, "user")
+                print(f"Welcome {name} to Pocket Vari Bank. Your Account number is, {user.account_no}.")
+                bank.account_list.append(user)
+            elif user_type == 'a':
+                user = Account(name, email, phone, password, "admin")
+                print(f"Welcome {name} to Pocket Vari Bank. Your Administrative Account number is, {user.account_no}.")
+                bank.account_list.append(user)
         else:
             print("Wrong Input! Try again.")
             
             
     elif op == 'l':
         user_type = input("User/Admin (u/a) : ")
-        name = input("Enter your name : ")
-        password = input("Password : ")
         
-        for user in bank.account_list:
-            if user.name == name:
-                if user.password == password:
-                    if user_type == 'a':
-                        print("Processing...")
-                        sleep(random.randint(3, 5))
-                        logged_in_as_admin()
+        if user_type == 'a' or user_type == 'u':
+            email = input("Enter your email : ")
+            password = input("Password : ")
+            
+            for user in bank.account_list:
+                if user.email == email:
+                    if user.password == password:
+                        if user_type == 'a':
+                            print("Processing...")
+                            sleep(random.randint(3, 5))
+                            logged_in_as_admin()
+                        else:
+                            print("Please wait...")
+                            sleep(random.randint(3, 5))
+                            logged_in_as_user()
                     else:
-                        print("Please wait...")
-                        sleep(random.randint(3, 5))
-                        logged_in_as_user()
+                        print("Sorry, Password didn't match. Try again.")
                 else:
-                    print("Sorry, Password didn't match. Try again.")
-            else:
-                print("Sorry, Name didn't match. Try again.")
+                    print("Sorry, email didn't match. Try again.")
+        else:
+            print("Wrong Input! Try again.")
                 
     elif op == 'e':
         print("Shutting down...")
